Server/requirements.txt
cryptography==36.0.2
cycler==0.11.0
Flask==2.0.3
frozenlist==1.3.0
gunicorn==20.1.0
httpcore==0.15.0
importlib-metadata==4.11.3
keyring==23.5.0
matplotlib==3.5.1
more-itertools==8.13.0
odict==1.8.1
optional-django==0.1.0
packaging==21.3
pathspec==0.9.0
pbkdf2==1.3
pefile==2017.9.3
Pillow==9.0.1
platformdirs==2.5.2
plumber==1.7
psutil==5.9.0
psycopg2==2.9.3
pyaes==1.6.1
pyasn1==0.4.8
pycparser==2.21
pycryptodome==3.14.1
pyerfa==2.0.0.1
pyinstaller==5.1
pyinstaller-hooks-contrib==2022.6
pyparsing==3.0.7
pystyle==2.0
pytesseract==0.3.9
python-dateutil==2.8.2
regex==2022.3.15
requests==2.26.0
rsa==4.8

